Improve collectd tool to handle additional unrecoverable tar error
There are cases when tar reports a different error string when it runs
out of space while tar'ing up a tarball ; "tar: error is not recoverable".

This update adds that string to the current list of strings that suggest
collect errored out due to space limitation.

Now that there are 3 error strings to look for, the block of script code that
checked for these error strings was re-factored to search a list of error
strings which makes the implementation cleaner and more maintainable.

If new error message strings are found in the future then such
strings need only be added to this new list.

Change-Id: Ifc677740c33372b48a4856ce00caea052580788d
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>
Signed-off-by: Scott Little <scott.little@windriver.com>