mellanox/diverse: update srpm paths in drivers
The srpm_path in these drivers points to a downloads directory. That
downloads directory has moved. This updates the path to point to the
current location.

Note: This will need to be updated again when the files in downloads
are moved and populated by the mirror scripts and reside in
/import/mirrors.

Signed-off-by: brian avery <brian.avery@intel.com>