Fix anaconda host IP caching to handle port number in URL
A hook was added to run before Anaconda installer to lookup the IP
address of the boot server and write it to /etc/hosts, in order to avoid
DNS queries during the installation. However, the regex to grab the
server from the kickstart URL did not handle the case where a port was
specified, ie http://bootserver:8080/..., and the hook would hang trying
to ping an invalid name.

This update extends the regex in the hook to ignore the port number, if
present.

Change-Id: Idec2219f0be55f08e7fd648a91a32428178715f6
Story: 2002856
Task: 22805
Signed-off-by: Jack Ding <jack.ding@windriver.com>
Signed-off-by: Scott Little <scott.little@windriver.com>