Enable Mellanox PMDs in ovs-dpdk build
The Mellanox devices are not currently supported by ovs-dpdk since they
are not built into the set of PMDs enabled.  Due to external build
dependencies, the default DPDK configuration sets them to disabled.

This update enables the mlx4 and mlx5 PMDs and adds the necessary build
dependencies to the openvswitch.spec file to properly build and link the
Mellanox PMDs.

Story: 2003104
Task: 23218

Change-Id: I9048b1b39e279261622f5add02c6642fab21e532
Signed-off-by: Matt Peters <matt.peters@windriver.com>