Fix controller swact error caused by drbd resizing
Fix a bug found in ‘provider/logical_volume/lvm.rb’ in method
‘def size=(new_size)’ with the variable
‘nuke_fs_on_resize_failure’. The conditional statement was always
returning true and thus wiping bytes at the start of the volume
even is the variable was set to false.

For example resizing the extension filesystem with the command
‘system controllerfs-modify’ would cause a dd command to be
executed and erase data. This was seen in the puppet.log

Story: 2002990
Task: 23004

Change-Id: I9ce4f9869d8b72549640d1a4181df02490451a88
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>