update-motd: add motd.head to /etc
The cron job for update-motd will run at the top of the hour and clobber
an original /etc/motd file provided by 'setup' RPM.

Consider that update-motd is intended to replace the motd provided by
setup, and provide that same content in the update-motd package.  In
this way, the original content from setup RPM will be displayed until
the cron job for update-motd runs.  After that, unless the user has
changed to the motd, the same content will be presented.

Partial-Bug: 1790863

Change-Id: I2b66ce7475525920c52e129102fa859fb29e0eed
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>