refactor audit
Package audit-config is created to config customized config
file of audit. Since there is no other change for audit,
we could replace srpm with rpm directly.

audit-config is set to depends on audit, so audit rpm will be
installed automatically.

Test:
Pass build and multi node deploy test. Confirm syslog.conf is
the same as before in the deploy.

Story: 2003768
Task: 27602
Depends-On: https://review.openstack.org/617174

Change-Id: I6101142642dd21c35e7db1352cc8c9aa05fba923
Signed-off-by: slin14 <shuicheng.lin@intel.com>