remove create-var-run-resource-agents.patch
The /var/run is on tmpfs, so the dir will be lost when reboot. This
issue is already fixed in upstream since resource-agents-3.9.5-116
with bz1462802-systemd-tmpfiles.patch. Our current version is 3.9.5-124.
So we could drop our patch safely.
Here is the upstream issue track link:
https://bugzilla.redhat.com/show_bug.cgi?id=1462802

Test:
Pass build and multi node deploy test.

Story: 2004019
Task: 27010

Change-Id: Ieb8ecd863d90fa0f08195f8765daf551127c8a1d
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>