Remove last patch of iscsi-initiator-utils
As we see in the patch, it changes %dir to %ghost to avoid RPM audit.
If we move the config file mod change to config package and use RPM
instead of SRPM, we have no audit issue anymore and can ignore related
change.
Deployment test pass and related file check pass!

Story: 2003768
Task: 28459
Depends-on: https://review.openstack.org/#/c/624584/

Change-Id: Ic23ccd740520e1942b3118a84cb03aef5f388332
Signed-off-by: zhipengl <zhipengs.liu@intel.com>