[Enhancement] Add system active alarms in collect logs
Currently the collect tool does not collect the active alarm list.
i.e. it doesn't contain the output of "fm alarm-list"

Although fm-event.log contains the history of events/alarms,
it is quite time-consuming to vet the fm-event.log for which alarms
were really active at the time of log collection.
A dump of alarm-list to the collect log will
greatly reduce this effort.

Story: 2004478
Tasks: 28193

Change-Id: I663fb055c244b31a84e682481f94bff82f610b90
Signed-off-by: Kwan, Louie <Louie.Kwan@windriver.com>