Scope parameter in ifup-aliases network script
The management interface uses the loopback interface for AIO-SX.
For the containerized deployment, the cluster-host interface shares
the same interface with management by default. It would need to be
scoped to host instead of global as it is an alias to the loopback
interface.
The scope parameter is added for the ifup-aliases to set scope
to host if the parent device is the loopback interface.

Story: 2004273
Task: 28590

Change-Id: Iee62c825ed1650b6cfdb0d4e5eb43b4a635e819a
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>