Disable mellanox driver support in DPDK for CentOS 7.6
This is a temp solution to make build pass with CentOS 7.6 kernel.
To enable mellanox driver support again, DPDK need be upgraded to 18.11
version, but openvswitch for DPDK 18.11 is not released yet.
And mellanox driver itself need be upgraded to 4.5-1.0.1.0 or later, to
support CentOS 7.6 and align with DPDK 18.11.

Depends-On: https://review.openstack.org/625785
Story: 2004521
Task: 28688

Change-Id: Ifee1956fbbe8d4cc0bf2f796fe265aeac840184c
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>