Configurable Host HTTP/HTTPS Port Binding
In order to avoid conflicts with containerized services
binding to standard HTTP (80) / HTTPS (443) port numbers,
the default port numbers are changed to 8080 and 8443.

Lighttpd port configuration is performed through puppet,
and the packaged lighttpd.conf uses port 80. As a result,
lighttpd is bind to port 80 before running config_controller.
This prevents patching before running config_controller.
This update changes the default http port to 8080 in the
packaged lighttpd.conf.

8080 is http port and 8008 is horizon port. The default
config file is changed here to be consistent with the port
number configured via puppet.

Story: 2004642
Task: 29300
Depends-On: https://review.openstack.org/#/c/634237/

Change-Id: I52b8f602dc2349ffabd9b90344dfafaf703ee4d7
Signed-off-by: Tao Liu <tao.liu@windriver.com>