Reduce the collectd samples retention period
Collectd creates a samples database within the
InfluxDB database which is stored in the rootfs.

The current 4 week retention period is too long
for larger systems and could lead to the rootfs
filling up.

This update reduces that retention perid to 1 week
to protect the rootfs from being filled up with
sample data until the samples database is moved
to a more appropriate location.

Change-Id: Ic59712849fa228f19d15919594d23edc43109a0b
Closes-Bug: 1827301
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>