Fix the helm-upload failure on controller-1
The user "www" is used to upload charts by using
helm-upload cmd. The root cause for the helm-upload
failure is the uid and gid of user "www" do not match
the uid and gid on controller-0.

Currently, www is created on controller-0 during ansible
configuration and the uid is set to "1877" explicitly.
On controller-1, it's created in puppet manifest but
it doesn't set the uid.

This commit is to add the user "www" in setup package so
that it's created after each node installation and it
makes sure the uid and gid are consistent on each node.

Change-Id: I59f2b379eedc4edf206916798d9837f97917ef7a
Closes-Bug: 1833603
Signed-off-by: Angie Wang <angie.wang@windriver.com>