Use ntpq refid to tell if peer controller reaches reliable time source
This is a partial fix only for ipv4.

The ntpq.py verify if a valid source is the reference of
peer controller when the peer controller is selected as
time server.
This change will avoid raising false alarm when a
controller uses peer controller as time server while
the peer uses a reliable time source (e.g, external time
server, or accurate time device).

Partial-Bug: 1834071

Change-Id: I9140e14b79cb09088c8061a06fae22df97526a70
Signed-off-by: Bin Qian <bin.qian@windriver.com>