Alarm is not raised in case PTP GM is lost
"No lock" PTP alarm is raised only when GM is and was not present
in a network. Current logic only reaises this alarm in case MAC
address of GM is the same as local MAC address. But it is only
the case when no external GM ever appeared in a PTP setup.
In case GM was present in a network and then lost we need to check
port status instead. PTP MAC address still points to an external GM.
But port status is changed from SLAVE to LISTENING state.

Change-Id: I30365685e6f44566702cc82534ab6ebf0613a731
Closes-bug: 1836884
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>