Add new ntp package to build
This is required to fix a bug with ntpq and IPV6 addresses. The ntpq
command truncates the remote addresses to 15 characters. This is not
long enough for IPV6 addresses. This has been fixed in version 4.2.8
which is not yet released by Centos. Patch
Fix-ntpq-truncates-IPV6-addresses.patch provides a subset of the fix.

https://github.com/ntp-project/ntp/commit/aeb3ee65bcd00a4fe5e3d028a45ae8a3fabe134f
https://bugs.ntp.org/show_bug.cgi?id=1128

Depends-On: https://review.opendev.org/680105
Partial-Bug: 1840687

Change-Id: If9d07acf913ebebead5505d44129f0644511b748
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>