Uprev SR-IOV CNI and device plugin image base
Currently, StarlingX uses a version of the SR-IOV CNI and device
plugin container images that are based on a certain commit reference.
This is done to ensure reliable and predicable behaviour until the
images can be locked down on a stable release version.

It is desirable to move to a later version of these images for
a couple of reasons (aside from bug fixes, etc):

1. The SR-IOV CNI image now uses an alpine base, rather than
   a Redhat base.
2. The SR-IOV device plugin allows a DPDK enabled pod with
   Mellanox NICs to run unprivileged.

This commit moves the image base forward.

Testing has been performed with netdevice and DPDK based
pod applications with various combinations of the following
devices:

Mellanox MT27700 Family [ConnectX-4]
Intel 82599ES 10-Gigabit SFI/SFP+ Network Connection
Intel Ethernet Controller X710 for 10GbE SFP+

Change-Id: Ia74e135b3e4b1a00465d4a8fd0b4650efdcfd2c5
Closes-Bug: 1843963
Closes-Bug: 1835020
Signed-off-by: Steven Webster <steven.webster@windriver.com>