Fix missing IP of alias interface
The ifup-aliases script assumes that the IPv4 address is always
defined. If the configuration is only for IPv6, the script would
generate an error and not process the IPv6 address of the interface.
This commit is to bring up the IPv6 interface even if the IPv4 address
is not defined.

Partial-Bug: 1834234

Change-Id: Ib0c4cbc7ec19cc0c0c485e4ad63c380aa8a49a4c
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>