De-couple kubernetes rpms from one another
The spec file from Centos was written to tightly couple
the kubernetes rpms to the same version and therefore prevent
any of them from being changed without installing the entire set.

This blocks the kubernetes upgrade procedure, which expects
components such as kubeadm or kubelet to up-version independently
from the other kubernetes components.

Refer to Upgrading Control Planes section of:
https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/kubeadm-upgrade/

In addition, hyperkube (unused) was packaged in multiple
rpms which is an rpm patching semantic violation.

Story: 2005860
Task: 36956
Change-Id: I26b7bc4b232635ac5f58aa9db79fcfe505c85fdc
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>