Drop initscripts patch running ipv6 dhcp as daemon
This commit rebases initscripts patch set, dropping
run-dhclient-as-daemon-for-ipv6.patch

Currently, ifup-eth tries running ipv6 dhclient with the one-shot
option, and if that fails, then retries indefinitely in the background.
That has the side-effect of causing the ifup-post script to not be run
if the first dhclient attempt fails, which will prevent routes on that
interface from being created.  This is especially problematic in the
case of a DOR, where the compute nodes may come up before dnsmasq is up
on the controller.
This is different from upstream centos, which will only try running
dhclient with the one-shot option for ipv6.
By reverting the initscripts patch to run as a daemon, ipv6 dhclient now
runs as one-shot only, and if it fails, ifup-eth script exits without
getting an address, and then the node fails to come up and reboot.
While this may result in the compute node having an extra reboot in a DOR,
that is preferable to the compute coming up incorrectly and requiring a
lock/unlock to recover.

Closes-bug: 1844579
Change-Id: I5b7f6b7c878dc4e4737d986f11fae3301585fb1c
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>