Package /etc/default/lldpd as a config file
The lldpd package currently does not package the /etc/default/lldpd
file as a config file, but it is modified at runtime by a puppet
manifest. As a result, if the lldpd package is updated on a system, it
would overwrite the modified file with the version from the package.

This update adds the %config(noreplace) to lldpd.spec for this file.

Change-Id: I82e62bdcac9ea07a3eaea0dfca5b1037b4b392d6
Partial-Bug: 1850695
Signed-off-by: Don Penney <don.penney@windriver.com>