CentOS 8: Upgrade python-requests to version 2.20.0
(1)Release Version Upgrade
(2)Matching code changes with el7 to el8

This package actually comes from the openstack package repo [0]
and the correct version is 2.21.0-3.
Since the CentOS folks have not created a cloud repo yet.
Ultimately this will need to be a python3 version.
We will need to rename the package to python3-requests.

[0] http://vault.centos.org/7.7.1908/cloud/Source/openstack-stein/

Story: 2006729
Task: 37659
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Change-Id: I8544995320fa440074554c6fdf0e1143bf68b582
Signed-off-by: Dongqi Chen <chen.dq@neusoft.com>