CentOS 8: Upgrade intel-ixgbe to version 5.6.3
Fix spec file for centos 8:

(1) Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(2)Upgrade tarball version from 5.5.5 to 5.6.3,
because the old version is missing support to build against CentOS 8.

Change-Id: I0e5a85f2daf18d39a528515797e9c1122660b526
Story: 2006728
Task: 37661
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>