CentOS 8: Upgrade drbd to version 9.0.21-1
Fix spec file for centos 8:

(1) Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(2) Upgrade tarball version from 8.4.11-1 to 9.0.21-1,
because the old version is missing support to build against CentOS 8.

(3) By the context of the function,
the patch 0001-remove_bind_before_connect_error
need not be retained.

Patch path is removed from the build_srpm.data file,
because there is no patch now.

Change-Id: I365ca2d1cc01bd56c56dd6f06260523f4597be54
Story: 2006728
Task: 37717
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>