CentOS 8: Upgrade mlnx-ofa_kernel to version 4.7.1.0.0.1
Fix spec file for centos 8:

(1) Add two BuildRequires: elfutils-libelf-devel, python3
because the elfutils-libelf-devel package contains the
libraries to create applications for handling compiled objects.

(2) Add a new global CENTOS8 for support centos8.
Put changed code into its own meta patch file
0002-Support-Centos8-OS.patch.

(3) Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(4) Upgrade srpm version from 4.5.1.0.1.1 to 4.7.1.0.0.1,
because the old version is missing support to build against CentOS 8.

Change-Id: I4591f291b09cf5a294cf60c2aa7e2118218b274d
Story: 2006728
Task: 37718
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>