CentOS 8: Upgrade tpmdd to version 4.18
Fix spec file for centos 8:

(1)Add two BuildRequires: elfutils-libelf-devel
because the elfutils-libelf-devel package contains the
libraries to create applications for handling compiled objects.

(2)Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(3)The following patches are dropped because the upstream has fixed.

UPSTREAM-0001-tpm-replace-msleep-with-usleep_range.patch
UPSTREAM-0002-tpm-reduce-tpm-polling-delay-in-tpm_tis_core.patch
UPSTREAM-0003-tpm-use-tpm_msleep-value-as-max-delay.patch
UPSTREAM-0004-tpm-wait-for-stat-to-specify-variable-polling-time.patch
UPSTREAM-0006-tpm-use-struct-tpm_chip.patch

(4)Upgrade tarball version from 4.12 to 4.18,
because the old version is missing support to build against CentOS 8.

Change-Id: Ie88f3f9d580aa22c42d68d12c71865dfc0daedad
Story: 2006728
Task: 37739
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>