CentOS 8: Upgrade integrity to version 4.18
Fix spec file for centos 8:

(1)Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(2)Upgrade tarball version from 4.12 to 4.18,
because the old version is missing support to build against CentOS 8.

(3)The following patch is dropped because the upstream has fixed.
Changes-for-CentOS-7.6-support.patch

Change-Id: I7d89c1b2a5a0d2022effa7a4da98aa83219d51d9
Story: 2006728
Task: 37740
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>