Remove obsolete patches for eventlet and requests
We are currently carrying patches for python-eventlet and
python-requests that are over three years old and were
created to "fix" a nova issue. Now that nova is in a
container, it no longer uses our eventlet or requests so
these patches do not have any effect.

The patches need to be removed to allow eventlet to be
upversioned, which is being done in another commit.

Change-Id: I730ba86d8f6ec8b88b060ada324cfd268af24b35
Story: 2006781
Task: 37583
Depends-On: https://review.opendev.org/#/c/700071/
Depends-On: https://review.opendev.org/#/c/700054/
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>