Mellanox Driver: Disable use of kernel page pool functionality
Problem:  The out-of-tree Mellanox driver detects the presence
of page pool support in the new kernel, and thus wants to use it.
However, page pools are not configured to be on in the new kernel
by default (CONFIG_PAGE_POOL), and not only that, the config
option is hidden ie. it is not user selectable. The built-in
Mellanox driver selects it, but we don't use the built-in driver.
The out-of-tree driver does compile but not all pieces of it
will load properly, specifically the mlx5 pieces which rely on
page pool functionality being enabled in the kernel.

Solution: Simply disable kernel page pool use in the
out-of-tree Mellanox driver, making it work the same way as
it did with the older kernel.

Change-Id: If7e7155867d539352fcd0ea3acd5a17dd9d9579f
Closes-Bug: 1860347
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>
(cherry picked from commit 424ba94a9aa9e64fe1c0d2099b63e7d979b492cd)