Fix containerd build failure
The 20200205T023000Z CENGN build failed on containerd due to a build
ordering issue. In the failed build, containerd was built ahead of
rpm, and the mock build environment for the containerd build ran with
the stock CentOS version of RPM. Unfortunately, it appears this
version of RPM fails when trying to build the debuginfo for a golang
package. There are currently two other golang packages in StarlingX,
but these have debuginfo disabled in the spec.

Adding a version-specific dependency in the containerd spec to ensure
the newer RPM is installed resolves the issue.

Change-Id: Ia7c85751012bbd0c3b83a2496bd7424e123eef93
Closes-Bug: 1862038
Co-Authored-By: Scott Little <scott.little@windriver.com>
Signed-off-by: Don Penney <don.penney@windriver.com>