Create Docker image for running Intel N3000 FPGA tools
This creates a Docker image for the purpose of running the tools needed
for programming the N3000 FPGA and querying information from it.

The expectation is that this will be run from sysinv, similar to how the
airship-armada tools are run.

The basic usage of these tools are covered in the following document
from Intel:
https://www.intel.com/content/dam/www/programmable/us/en/pdfs/literature/ug/ug-ias-n3000.pdf

Change-Id: I94a7fb9fe6b348ced014666b37849691f72766e8
Story: 2006740
Task:  38829
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>