Kernel: Workaround broken bios affecting iommu
Problem:
Broken bios creates inaccurate DMAR tables,
reporting some bridges as having endpoint types.
This causes IOMMU initialization to bail
out early with an error code, the result of
which is vfio not working correctly.
This is seen on some Skylake based Wolfpass
server platforms with up-to-date bios installed.

Solution:
Instead of just bailing out of IOMMU
initialization when such a condition is found,
we report it and continue.  The IOMMU ends
up successfully initialized anyway.  We do this
only on platforms that have the Skylake bridges
where this issue has been seen.

This change is inspired by a similar one posted by
Lu Baolu of Intel Corp to lkml

https://lkml.org/lkml/2019/12/24/15

Change-Id: Ief2df7099b6118eab7f99d5531616926a7a3eb27
Closes-Bug: 1847335
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>