Upversion kubernetes to v1.18.1
Upgrade to kubernetes v1.18.1 and include a temporary patch for the
current StarlngX behavior related to hugepages.

Based on the current StarlingX configuration, container hugepages are
mounted with the following option that already contains an 'i' as a
suffix: pagesize=1Gi.

This temporary patch updates the logic associated with
https://github.com/kubernetes/kubernetes/commit/03ecc20 to avoid
appending an additional 'i' at the end of the size string. This extra
'i' is not handled by ParseQuantity() and results is a pod stuck
Terminating as the mount is not removed from the container.

The new multiple hugepage feature will be enabled an evaluated with the
intent of upstreaming this (or a comparable fix) OR adjusting StarlingX
hugepage behavior to align with the expected upstream behavior and
dropping this patch.

In addition:
- Update kubelet-cgroup-setup.sh, as now pids and hugetlb both appear to
  get auto-deleted on reboot so make sure these are recreated.
- Update spec to remove hyperkube dependencies as each component is now
  built independently and needs to be install individually.

Change-Id: I34263fa33c61c7a5bce1002a4f5edf5b850aa189
Story: 2006999
Task: 37770
Depends-On: https://review.opendev.org/#/c/718370/
Depends-On: https://review.opendev.org/#/c/718566/
Signed-off-by: Robert Church <robert.church@windriver.com>