Add kubernetes support for isolated cpus
This introduces the concept of "isolated CPUs", which are CPUs that
have been isolated at the kernel level via the "isolcpus" kernel boot
parameter.

When starting the kubelet process, the set of reserved CPUs (including
both platform and isolated CPUs) will be specified via
'--reserved-cpus'.  The isolated CPUs will be identified by looking at
"/sys/devices/system/cpu/isolated" and treated separately from the
platform CPUs (which are used to run infrastructure pods).

A plugin (outside the scope of this commit) exposes the isolated
CPUs to kubelet via the device plugin API.

If a pod specifies some number of "isolcpus" resources, the device manager
will allocate them.  In this code we check whether such resources have
been allocated, and if so we set the container cpuset to the isolated
CPUs.  This does mean that it really only makes sense to specify "isolcpus"
resources for best-effort or burstable pods, not for guaranteed ones since
that would throw off the accounting code.  In order to ensure the accounting
still works as designed, if "isolcpus" are specified for guaranteed pods,
the affinity will be set to the non-isolated CPUs.

Story: 2008760
Task: 42161
Change-Id: I7bd2eabb4c82faea63e3ad129ef735b9d1223e11
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>