Fix sriov device plugin image build
Previous commit d204f10ab5 introduced a build script to assist
in building the SR-IOV device plugin.

The script utilizes a Makefile to do build the plugin binary,
then the image.

Building the binary depends on go being present on the host. If it
is not, the build will fail.

Building the binary is actually not required, as it will be also
done in a container as part of the 'make image', rather than copying
the binary from the host.

Closes-Bug: #1878224
Change-Id: I4499ea2bbef4b3da8a154c69a07b415574517500
Signed-off-by: Steven Webster <steven.webster@windriver.com>