switch to containerd.service file from upstream
Back when the containerd package was first added to the build,
the designer who added it didn't realize that the upstream source
already contained a "containerd.service" file and so they added
a separate one.

It turns out that the upstream source *does* have a service file,
and it also contains some additional settings that we might want
to pick up.  Furthermore, there are additional changes in more
recent versions of the package.  As such, we want to switch to
use the service file from the upstream source instead of a custom
one.

The upstream service file wants to run /usr/local/bin/containerd
so we just make a symlink at that location pointing to the
current binary location.

Closes-Bug: 1884111
Change-Id: I5ed4f46a7bcceb0d0f71abb26590160fb62c0b7b
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>