CVE-2019-15916: kernel: memory leak in register_queue_kobjects
Upversion the kernel to CentOS 7.8 which is 3.10.0-1127.13.1 .

We get to drop the ipvs memory leak patch because it is now
incorporated.

On std, we remove the rcuc patch because it wasn't being
used and thus was just polluting the patch directory.

On rt, we also get rid of the robustify CFS bandwidth locking
patch because it is already incorporated.

Minor tweaks for compilation.

Change-Id: I5ceac098d2487c77138e3ee2d39e80b2e20b7af0
Partial-Bug: 1881429
Depends-On: https://review.opendev.org/#/c/739836
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>