kubeadm: create platform pods with zero CPU resources
We want to specify zero CPU resources when creating the manifests
for the static platform pods, as a workaround for the lack of
separate resource tracking for platform resources.

We also specify zero CPU resources for the coredns deployment.
manifests.go appears to be the main file for this, not sure if the
others are used by I changed them just in case.

Story: 2008760
Task: 42163
Change-Id: I6410b8af556d5167d1813e7545fad8baa27b1100
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>