dhcp: fix CVE-2019-6470
Upgrade dhcp pkg to dhcp-4.2.5-82.el7.centos.src.rpm

Adjust the context of the patch to match to apply the new version.
At the same time as the new version depends on the bind-export
pacakges, so we also add the dependence package in tools repo.
 bind-export-libs-9.11.4-26.P2.el7.x86_64.rpm
 bind-export-devel-9.11.4-26.P2.el7.x86_64.rpm

In addition, since the patch dhcp-dhclient_ipv6_prefix.patch set the
default prefixlen to 128, which is usually the specifications call
for host address and it doesn't include any on-link information.
By contrast, 64 indicates that's subnet area, and this vaule is used
frequently as usual. So we still use the previous value 64.
As a result we don't need to modify the relevant place where every
application code needed for the compatibility any more.

Depends-On: https://review.opendev.org/c/starlingx/tools/+/772241

Story: 2008532
Task: 41638
Change-Id: I0305711790d8e3fb1adfa69e1077468456b65d84
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>