Revert "dhcp: fix CVE-2019-6470"
This reverts commit 613fbf258f72042f912a1fde5608168b1068db36.

Since this upversioned package updates the prefixlen to 128, and it
will occur all hosts offline after booting off the controller-0.
At the same time this issue will block the use of recent loads for
both development and test activities. So we revert the patch firstly,
and investigate deeply then send the new review and request of the
upgraded patch with the appropriate offline fix.

Closes-Bug: #1915050

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: I02ecaa1bda463efb38d9c32a47f2221d0de7f99d