Handle labeled pods after stabilized
Pods that are in a k8s deployment, daemonset, etc can be labeled as
restart-on-reboot="true", which will automatically cause them to be
restarted after the worker manifest has completed in an AIO system.
It may happen, however, that k8s-pod-recovery service is started
before the pods are scheduled and created at the node the script is
running on, causing them to be not restarted. The proposed solution is
to wait for stabilization of labeled pods before restarting them.

Closes-Bug: 1900920
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I5c73bd838ab2be070bd40bea9e315dcf3852e47f