Add alternative command to disable lldp agent for i40e devices
LLDP information is not available for certain i40e network devices when
running system host-lldp-neighbor-show.

This is caused by the firmware lldp agent on the devices not getting
disabled by the i40e-lldp-configure.sh script which is invoked by lldpd.

The command used to disable the firmware lldp agent in the script works
for some firmware versions found on devices, but not others. This change
adds an ethtool command to disable the lldp agent which works for these
other firmware versions.

From testing, the ethtool method is used for firmware versions 5.05 and
8.10. The sysfs method is used for firmware version 7.10. In all cases,
the driver version is 2.14.13

Closes-Bug: 1923665

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Ifac34091599bd4020bf55cc1b8ba3119edccb297
(cherry picked from commit de263f633ea359507357d3d4c53e98a71bff5afc)