Armada: Fix tiller stuck connecting to postgres database
Tiller may start executing before IPv6 network is fully initialized.
This will result in tiller not being fully functional.
The liveness probe will detect that tiller didn't start properly and
restart it. But this might happen an unlimited number of times in a row.

Wait until ping is succesful to the ip of the postgres database.
This ensures that networking finished setting up.
Credits to Cole Walker <cole.walker@windriver.com> for proposing the
idea.

Depends-On: I177bb628497611eb64472291a04d635856c26590
Closes-Bug: 1928141
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I9c5be3f30fad2650e6aa53fb80ef44f7798813ed