Update the liveness probe to verify postgres connectivity
Change the tillerLivenessProbeTemplate to test the connectivity to the
postgres backend. We will override the periodSeconds and
failureThreshold when installing the helm chart to trigger a restart of
the tiller pod over a swact when the postgres DB/server moves from one
controller to the other.

This will help guarantee that the tiller connection is always
re-established if the connectivity to the postgres backend fails.

Change-Id: I7fbed33a8c821f6c9254f58d5953e2115cf4141a
Related-Bug: #1917308
Signed-off-by: Robert Church <robert.church@windriver.com>
(cherry picked from commit b3540ccfdfa6956fb20c62e5e5bb76af56d2ab63)