Revert "Remove recover operations to "restart-on-reboot" pods"
This reverts commit 8abcbf6fb1951b25e9964933558b75b9aff88135.

Reason for revert:

After performing a backup and restore on an AIO-SX system, SRIOV pods do
not return to a running state and are instead stuck in "container
creating". The workaround for this is to restart SRIOV pods when the
system unlocks.

Reverting this commit to allow users to label SRIOV pods and have them
restarted by k8s-pod-recovery. Labelled pods will be restarted by
k8s-pod-recovery and will be running after backup and restore is
completed.

This change has been tested by performing backup and restore on an
AIO-SX system. SRIOV pods now come up correctly when labelled with
restart-on-reboot=true

Closes-Bug: 1928965

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I9c520c0a47aabca7b96e50adf0f71742f4199c2f