Wait for SRIOV device plugin before recovering labeled pods
This change modifies the k8s-pod-recovery service to wait for the
kube-sriov-device-plugin-amd64 pod on the local node to become
available before proceeding with the recovery of
restart-on-reboot=true labeled pods.

This is required because of a race condition where pods marked for
recovery would be restarted before the device plugin was ready and
the pods would then be stuck in "ContainerCreating".

The fix in this commit uses the kubectl wait ...
command to wait for the daemonset to be available. A timeout of 360s
has been set for this command in order to all enough time on busy
systems for the device-plugin pod to come up. The wait command
completes as soon as the pod is ready.

Closes-Bug: 1928965

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Ie1937cf0612827b28762049e2dc440e55726d4f3