Upgrade pf-bb-config to 21.6 and new config file
After some reports of Mt.Bryce hang with the current version (21.3) of
pf-bb-config and from Intel's feedback this utility is being upgraded
to latest version tagged as 21.6.

Also, another configuration file is taken from the official repo at
GitHub, which is supposed to fit better with the new version and the
current device usage requirements.

Partial-Bug: #1933372
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: If175cc5b33b59fa2095d1b466dcbc8b724287817