Worker node slow to become online after power out
During a DOR (Dead Office Recovery) procedure, the worker nodes are
taking longer than expected to become online, if the cluster network
is IPv6. This is related to an extra reboot occurring during the
bring-up process. Today only one DHCPv6 attempt is done, and, as the
controllers are still booting up, no response is available.
Eventually, this lack of a management IPv6 address will lead to an
extra reboot.

The correction is to make the DHCPv6 client continuously attempt to
get a management IP address from the server. If the management
network is IPv4, the DHCPv6 client is not called and will not
trigger continuous DHCPv6 requests.

Closes-bug: 1934676

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ife5b06c5294e0014918073e9f6ee8bd372b85a78