Add pf-bb-config configuration file for N3000
In the context of starting usage of pf-bb-config tool to configure the
N3000 (Intel FPGA) FEC device, it is necessary to include in the utility
package the corresponding file with the parameters set into the device.

For now, it is only fpga_5gnr/fpga_5gnr_config_vf.cfg from pf-bb-config
GitHub site (https://github.com/intel/pf-bb-config/tree/v21.6) that is
being taken for N3000 configuration, and used as is.

Task: 42784
Story: 2009036
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I219ddcf639f0ac80644b6f25cd50b883752c012e