Patch pf-bb-config to return -1 on failed config
Current implementation of pf-bb-config utility (tag 21.6) does not check
the result of configure_device() function. Thus, if pf-bb-config is
running in a script or automated procedure (like when applying a puppet
manifest), any configuration failure cannot be detected unless the
output is parsed.
This change patches the source code of pf-bb-config in the same way that
was proposed by the PR: https://github.com/intel/pf-bb-config/pull/3. If
that change is merged, then in the future when pf-bb-config is upgraded
the patch can be removed.

Task: 42877
Story: 2009036
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: Ic0ec0c7b9088a573cadf468839be1c4445d7abb2