Modify go package to be compatible with golang 1.16.6
Most packages will compile using the new default golang version (1.16.6).
For some packages it is a transparent change, but those reliant on
the old GOPATH method require the addition of  ...

    go env -w GO111MODULE=auto

... to their build instruction to restore GOPATH compatibility.

For Kubernetes version 1.18.1, I explicitly set to compile using
golang version 1.13.9 as this is the version recommended by Kubernetes.

Story: 2008972
Task: 42655
Depends-On: https://review.opendev.org/c/starlingx/compile/+/804123
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I0d813cd245cb46932bf2096b641c4972f8a49efc